WordFrequency.py
Right now the decades & keywords are hardcoded in, I'll add I/O to just specify keywords and dynamically create the decades list this week. You also need to manually type in the filepath to the Danish Json files in the script, I'll also add the I/O for that this week.