Overwrite & optional filtering
If the output directory already exists, it will now be overwritten. Also added an optional argument ("-f") which will filter the raw text by removing stop-words and unnecessary characters, returning an array of individual words. I just added it to make the word frequency scripts run faster, because it takes a really long time to build the array of cleaned text for the frequency distribution. Note that if you specify the filtering argument, the script will take ~1 hour to run.