Update WordFrequency.py
Added max/min/avg functionality as well as basic word frequency as a percentage of total words. Only one of these can be graphed per run of the script, but all of them are outputted in a text file at the end of each run.
Also simplified/consolidated/cleaned the code a lot.