Removed -n argument
Removed -n so now it's just -num for top words for each decade. I initially had -n as a boolean flag followed by -num as the number of words the user wanted, but then figured it would be simpler to just have -num and replace the "if -n:" lines with exception blocks.